iphone: fix spec on device

+++ spec/framework_spec/app/spec/core/signal/trap_spec.rb
@@ -15,14 +15,14 @@ describe "Signal.trap" do
   it "returns the previous handler" do
     Signal.trap(:HUP, @saved_trap).should equal(@proc)
   end
-
+if System.get_property('platform') != 'APPLE'
   it "accepts a block in place of a proc/command argument" do
     Signal.trap(:HUP) { ScratchPad.record :block_trap }
     Process.kill :HUP, Process.pid
     sleep 0.5
     ScratchPad.recorded.should == :block_trap
   end
-
+end
   ruby_version_is ""..."1.9" do
     it "ignores the signal when passed nil" do
       Signal.trap :HUP, nil
